remove e2e tests. Replace some TinyTest scripts with mocha
We have found the creation and maintenance of our e2e scripts to be
extremely expensive and time consuming and ultimately producing little
value to be worth it for now. While it would be great to focus on
testing behavior instead of implementation, the flakiness of our builds
and the amount of time required to maintain these tests has proven to be
not worth it. We are removing the e2e tests for now. The scripts and
specs will remain for a while in case we want to revive this effort.

Also, a subset of the TinyTest suite for meteor package testing has been
replaced with nodejs Mocha testing. These tests can be run extremely
fast as they do not rely on the Meteor environment spinning up. In
general, it would be great if most of our testing focused in this area.

We still have tiny test specs around that we plan on using for
integration testing. (These are integration specs because they rely on
the Meteor environment to spin up).